Addresses minor comments
