+++ Example/Auth/Tests/FIRAuthTests.m
@@ -23,6 +23,7 @@
 #import "FirebaseCommunity/FIRGoogleAuthProvider.h"
 #import "FirebaseCommunity/FIRAdditionalUserInfo.h"
 #import "FIRAuth_Internal.h"
+#import "FIRAuthOperation.h"
 #import "FIRAuthErrorUtils.h"
 #import "FIRAuthDispatcher.h"
 #import "FIRAuthGlobalWorkQueue.h"
@@ -423,7 +424,8 @@ static const NSTimeInterval kWaitInterval = .5;
                        FIRVerifyPhoneNumberResponseCallback callback) {
     XCTAssertEqualObjects(request.verificationCode, kVerificationCode);
     XCTAssertEqualObjects(request.verificationID, kVerificationID);
-    XCTAssertEqualObjects(request.operation, operationType(FIRAuthOperationTypeSignUpOrSignIn));
+    XCTAssertEqualObjects(request.operation,
+                          FIRAuthOperationString(FIRAuthOperationTypeSignUpOrSignIn));
     dispatch_async(FIRAuthGlobalWorkQueue(), ^() {
       id mockVerifyPhoneResponse = OCMClassMock([FIRVerifyPhoneNumberResponse class]);
       [self stubTokensWithMockResponse:mockVerifyPhoneResponse];
