+++ Example/Auth/Tests/FIRUserTests.m
@@ -27,6 +27,7 @@
 #import "FIRAuthErrorUtils.h"
 #import "FIRAuthBackend.h"
 #import "FIRAuthGlobalWorkQueue.h"
+#import "FIRAuthOperation.h"
 #import "FIRGetAccountInfoRequest.h"
 #import "FIRGetAccountInfoResponse.h"
 #import "FIRSetAccountInfoRequest.h"
@@ -1882,9 +1883,9 @@ static const NSTimeInterval kExpectationTimeout = 1;
     XCTAssertEqualObjects(request.verificationID, kVerificationID);
     XCTAssertEqualObjects(request.verificationCode, kVerificationCode);
     if (isLinkOperation) {
-      XCTAssertEqualObjects(request.operation, operationType(FIRAuthOperationTypeLink));
+      XCTAssertEqualObjects(request.operation, FIRAuthOperationString(FIRAuthOperationTypeLink));
     } else {
-      XCTAssertEqualObjects(request.operation, operationType(FIRAuthOperationTypeUpdate));
+      XCTAssertEqualObjects(request.operation, FIRAuthOperationString(FIRAuthOperationTypeUpdate));
     }
     XCTAssertEqualObjects(request.accessToken, kAccessToken);
     dispatch_async(FIRAuthGlobalWorkQueue(), ^() {
