+++ Firebase/Auth/Source/FIRUser.m
@@ -27,6 +27,7 @@
 #import "FIRAuthErrorUtils.h"
 #import "FIRAuthGlobalWorkQueue.h"
 #import "FIRAuthSerialTaskQueue.h"
+#import "FIRAuthOperation.h"
 #import "FIRAuth_Internal.h"
 #import "FIRSecureTokenService.h"
 #import "FIRUserInfoImpl.h"
@@ -635,8 +636,8 @@ static void callInMainThreadWithAuthDataResultAndError(
       completion(error);
       return;
     }
-    NSString *operation = isLinkOperation ? operationType(FIRAuthOperationTypeLink) :
-                                            operationType(FIRAuthOperationTypeUpdate);
+    NSString *operation = isLinkOperation ? FIRAuthOperationString(FIRAuthOperationTypeLink) :
+                                            FIRAuthOperationString(FIRAuthOperationTypeUpdate);
     FIRVerifyPhoneNumberRequest *request = [[FIRVerifyPhoneNumberRequest alloc]
         initWithVerificationID:phoneAuthCredential.verificationID
               verificationCode:phoneAuthCredential.verificationCode
